LPS-167730 Move SiteNavigationMenuPortletKeys to -api, prepare for next
