+++ modules/apps/site-navigation/site-navigation-api/src/main/java/com/liferay/site/navigation/constants/SiteNavigationMenuPortletKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.site.navigation.menu.web.internal.constants;
+package com.liferay.site.navigation.constants;
 
 /**
  * @author Eudaldo Alonso
